One username (#90)
* Duplicate Check

This code checks the database in order to prevent the code from attempting to register multiple users under the same username.

Had to remove the test case that checked for failure, since a 400 error no longer occurs when a double username is registered.

* Error Response

Sends a message to the user in the event that registration fails because the username is already in use.